apka/requirements2.txt
pygetwindow
pyautogui
opencv-python
ffmpeg-python
